Revert "fix(codex): preserve child hooks across turns (#128485)"
This reverts commit a5c8e00cb6701e89604bb96ba91583680c2836f6.
