+++ extensions/codex/src/app-server/native-subagent-monitor.test.ts
@@ -1645,68 +1645,6 @@ describe("CodexNativeSubagentMonitor", () => {
     client.close();
   });
 
-  it.each([
-    {
-      name: "the child remains actively claimed",
-      claim: true,
-      threadStatus: "active",
-      status: "inProgress",
-      expectedRenewals: 1,
-      expectedReleases: 0,
-    },
-    {
-      name: "the active child was never claimed",
-      claim: false,
-      threadStatus: "active",
-      status: "inProgress",
-      expectedRenewals: 0,
-      expectedReleases: 0,
-    },
-    {
-      name: "the claimed child is resumable",
-      claim: true,
-      threadStatus: "idle",
-      status: "interrupted",
-      expectedRenewals: 0,
-      expectedReleases: 1,
-    },
-  ] as const)("renews only when $name after an authoritative read", async (testCase) => {
-    const client = createClient();
-    client.setThreadRead(
-      "child-thread",
-      threadRead({ threadStatus: testCase.threadStatus, status: testCase.status }),
-    );
-    const releaseDirectChild = vi.fn();
-    const renewDirectChild = vi.fn();
-    const monitor = new CodexNativeSubagentMonitor(client as never, createRuntime());
-    const owner = monitor.registerParent({
-      parentThreadId: "parent-thread",
-      claimDirectChild: () => releaseDirectChild,
-      renewDirectChild,
-    });
-    owner.bindTurn("turn-1");
-    await notifyChildStarted(client);
-    if (testCase.claim) {
-      await client.notify({
-        method: "item/completed",
-        params: {
-          threadId: "parent-thread",
-          turnId: "turn-1",
-          item: directSpawnItem("v1", "parent-thread", "child-thread"),
-        },
-      });
-    }
-
-    await expect(monitor.reconcileChildThread("child-thread")).resolves.toBe(false);
-
-    expect(renewDirectChild).toHaveBeenCalledTimes(testCase.expectedRenewals);
-    expect(releaseDirectChild).toHaveBeenCalledTimes(testCase.expectedReleases);
-    if (testCase.expectedRenewals === 1) {
-      expect(renewDirectChild).toHaveBeenCalledWith("child-thread");
-    }
-    client.close();
-  });
-
   it("does not replay stale history while a system-error child still has an active turn", async () => {
     vi.useFakeTimers();
     try {
