+++ extensions/codex/src/app-server/run-attempt-resources.ts
@@ -194,10 +194,6 @@ export function prepareCodexAttemptResources(prompt: CodexAttemptPrompt) {
   };
   const registerNativeSubagentMonitor = (parentThreadId: string) => {
     unregisterNativeSubagentMonitor();
-    // Child recovery outlives this attempt's mutable state. Capture the relay
-    // that admitted the child so a successor foreground turn cannot renew or
-    // release the child under the successor's policy.
-    const nativeHookRelay = state.nativeHookRelay;
     state.nativeSubagentMonitor = codexNativeSubagentMonitorRuntime.register({
       client: state.client,
       parentThreadId,
@@ -207,16 +203,9 @@ export function prepareCodexAttemptResources(prompt: CodexAttemptPrompt) {
       retainClient: () => retainSharedCodexAppServerClientIfCurrent(state.client),
       retainParentThread: (protectedThreadId) =>
         protectCodexAppServerLiveThread(state.client, protectedThreadId),
-      ...(nativeHookRelay
-        ? {
-            claimDirectChild: (childThreadId: string) =>
-              nativeHookRelay.claimDirectChild(childThreadId),
-            renewDirectChild: (childThreadId: string) =>
-              nativeHookRelay.renewDirectChild(childThreadId),
-            rejectPendingDirectChild: (childThreadId: string, reason: string) =>
-              nativeHookRelay.rejectPendingDirectChild(childThreadId, reason),
-          }
-        : {}),
+      claimDirectChild: (childThreadId) => state.nativeHookRelay?.claimDirectChild(childThreadId),
+      rejectPendingDirectChild: (childThreadId, reason) =>
+        state.nativeHookRelay?.rejectPendingDirectChild(childThreadId, reason),
       ...(params.sessionKey && params.agentHarnessTaskRuntimeScope
         ? {
             onDirectChildAccepted: () => {
@@ -225,10 +214,6 @@ export function prepareCodexAttemptResources(prompt: CodexAttemptPrompt) {
           }
         : {}),
     });
-  };
-  const bindNativeTurn = (turnId: string) => {
-    state.nativeHookRelay?.bindForegroundTurn(turnId);
-    state.nativeSubagentMonitor?.bindTurn(turnId);
   };
   const releaseCurrentRoute = () => {
     state.detachRouteAbort();
@@ -245,12 +230,8 @@ export function prepareCodexAttemptResources(prompt: CodexAttemptPrompt) {
   const requesterChannel = params.messageChannel ?? params.messageProvider;
   const requester = buildCodexHookRequester(params);
   const buildNativeHookRelayFinalConfigPatch = (
-    decision:
-      | { action: "resume"; binding: CodexAppServerThreadBinding }
-      | { action: "start"; preserveExistingBinding: boolean },
+    decision: { action: "resume"; binding: CodexAppServerThreadBinding } | { action: "start" },
   ) => {
-    const composeWithExistingRoute =
-      decision.action === "resume" || decision.preserveExistingBinding;
     state.nativeHookRelay?.unregister();
     if (params.pluginHarnessToolPolicyRestricted === true) {
       state.nativeHookRelay = undefined;
@@ -261,7 +242,6 @@ export function prepareCodexAttemptResources(prompt: CodexAttemptPrompt) {
     }
     state.nativeHookRelay = createCodexNativeHookRelay({
       options: options.nativeHookRelay,
-      composeWithExistingRoute,
       generation:
         decision.action === "resume" ? decision.binding.nativeHookRelayGeneration : undefined,
       generationMismatchGraceMs:
@@ -302,9 +282,6 @@ export function prepareCodexAttemptResources(prompt: CodexAttemptPrompt) {
       },
     });
     return {
-      ...(composeWithExistingRoute && state.nativeHookRelay
-        ? { activateThreadBinding: state.nativeHookRelay.activateForegroundBinding }
-        : {}),
       configPatch: state.nativeHookRelay
         ? buildCodexNativeHookRelayConfig({
             relay: state.nativeHookRelay,
@@ -332,7 +309,6 @@ export function prepareCodexAttemptResources(prompt: CodexAttemptPrompt) {
     releaseSandboxExecEnvironment,
     runCleanupStep,
     registerNativeSubagentMonitor,
-    bindNativeTurn,
     releaseCurrentRoute,
     startupTimeoutMs,
     buildNativeHookRelayFinalConfigPatch,
