+++ extensions/codex/src/app-server/run-attempt-route.ts
@@ -19,7 +19,6 @@ export async function prepareCodexAttemptRoute(
     trajectoryRecorder,
     releaseCurrentRoute,
     registerNativeSubagentMonitor,
-    bindNativeTurn,
     activateNativePreToolUseFailureFallback,
     releaseSandboxExecEnvironment,
     releaseSharedClientLeaseOnce,
@@ -86,7 +85,6 @@ export async function prepareCodexAttemptRoute(
       }
       resourceState.detachRouteAbort = attachRouteAbort(resourceState.turnRoute);
       await resourceState.turnRoute.activate({
-        onTurnStarted: bindNativeTurn,
         onNotificationReceived: noteNotificationReceived,
         onNotification: enqueueNotification,
         onRequest: handleServerRequest,
