+++ src/agents/harness/native-hook-relay-state.ts
@@ -11,11 +11,8 @@ function getNativeHookRelaySharedState(): NativeHookRelaySharedState {
     relays: new Map(),
     relayBridges: new Map(),
     invocations: [],
-    approvalOwners: new WeakMap(),
     pendingPermissionApprovals: new Map(),
-    pendingPermissionApprovalOwners: new Map(),
     pendingPreToolUseApprovals: new Map(),
-    pendingPreToolUseApprovalOwners: new Map(),
     permissionApprovalWindows: new Map(),
     permissionAllowAlwaysApprovals: new Map(),
   };
