+++ src/agents/harness/native-hook-relay.ts
@@ -1,9 +1,20 @@
 /** Native harness hook event relay and public Plugin SDK facade. */
+import { randomUUID } from "node:crypto";
+import {
+  MAX_TIMER_TIMEOUT_MS,
+  resolveExpiresAtMsFromDurationMs,
+} from "@openclaw/normalization-core/number-coercion";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
+import { resolveOpenClawStateSqlitePath } from "../../state/openclaw-state-db.paths.js";
+import { retainBeforeToolCallForNativeHookRelay } from "./host-capability.js";
 import {
   clearNativeHookRelayBridgesForTests,
+  NATIVE_HOOK_BRIDGE_REPLACEMENT_RECORD_GRACE_MS,
   NATIVE_HOOK_RELAY_BRIDGE_STALE_REGISTRATION_ERROR,
   readNativeHookRelayBridgeRecordIfExists,
+  registerNativeHookRelayBridge,
+  renewNativeHookRelayBridgeRecord,
+  unregisterNativeHookRelayBridge,
   isRetryableNativeHookRelayBridgeLookupError,
 } from "./native-hook-relay-bridge.js";
 import {
@@ -12,23 +23,23 @@ import {
   normalizeNativeHookToolName,
   readNativeHookRelayApprovalMode,
 } from "./native-hook-relay-codec.js";
-import { processNativeHookRelayInvocation } from "./native-hook-relay-events.js";
 import {
-  getNativeHookRelayRegistrationForTests,
-  getNativeHookRelayRoute,
-  listNativeHookRelayRoutes,
-  pruneExpiredNativeHookRelays,
-  registerNativeHookRelayLifecycle,
-  registerRetainedNativeHookRelayLifecycle,
-  resolveNativeHookRelayInvocationBinding,
-  unregisterNativeHookRelay,
-} from "./native-hook-relay-lifecycle.js";
-import type { NativeHookRelayRetention } from "./native-hook-relay-lifecycle.js";
-import { formatPermissionApprovalDescription } from "./native-hook-relay-permission-gateway.js";
+  buildNativeHookRelayCommandWithStateDatabase,
+  resolveNativeHookRelayCommandTimeoutMs,
+} from "./native-hook-relay-command.js";
 import {
-  permissionRequestContentFingerprint,
-  permissionRequestToolInputKeyFingerprint,
+  nativeHookRelayEventHasLocalWork,
+  nativeHookRelayEventToolMatcher,
+  processNativeHookRelayInvocation,
+} from "./native-hook-relay-events.js";
+import {
   clearNativeHookRelayPermissionsForTests,
+  formatPermissionApprovalDescriptionForTests as formatPermissionApprovalDescriptionForTestsImpl,
+  permissionRequestContentFingerprintForTests as permissionRequestContentFingerprintForTestsImpl,
+  permissionRequestToolInputKeyFingerprintForTests as permissionRequestToolInputKeyFingerprintForTestsImpl,
+  pruneNativeHookRelayPermissionAllowAlways,
+  removeNativeHookRelayPermissionState,
+  removeNativeHookRelayPreToolUseApprovals,
   setNativeHookRelayDeferredToolApprovalRequesterForTests as setNativeHookRelayDeferredToolApprovalRequesterForTestsImpl,
   setNativeHookRelayPermissionApprovalRequesterForTests as setNativeHookRelayPermissionApprovalRequesterForTestsImpl,
 } from "./native-hook-relay-permissions.js";
@@ -43,13 +54,16 @@ import type {
   InvokeNativeHookRelayParams,
   NativeHookRelayEvent,
   NativeHookRelayInvocation,
+  NativeHookRelayPermissionApprovalRequest,
   NativeHookRelayPermissionApprovalRequester,
   NativeHookRelayProcessResponse,
   NativeHookRelayRegistration,
   RegisterNativeHookRelayParams,
 } from "./native-hook-relay-types.js";
+import { NATIVE_HOOK_RELAY_EVENTS } from "./native-hook-relay-types.js";
 import {
   isJsonValue,
+  normalizePositiveInteger,
   readNativeHookRelayEvent,
   readNativeHookRelayProvider,
   readNonEmptyString,
@@ -57,7 +71,6 @@ import {
 } from "./native-hook-relay-utils.js";
 export { buildNativeHookRelayCommand } from "./native-hook-relay-command.js";
 export { resolveNativeHookRelayDeferredToolApproval } from "./native-hook-relay-permissions.js";
-export type { NativeHookRelayRetention } from "./native-hook-relay-lifecycle.js";
 export type {
   NativeHookRelayEvent,
   NativeHookRelayProcessResponse,
@@ -65,22 +78,395 @@ export type {
   NativeHookRelayRegistrationHandle,
 } from "./native-hook-relay-types.js";
 
+const DEFAULT_RELAY_TTL_MS = 30 * 60 * 1000;
 const log = createSubsystemLogger("agents/harness/native-hook-relay");
-const { invocations } = nativeHookRelayState;
+
+const { relays, relayBridges, invocations } = nativeHookRelayState;
+type RelayLifetime = {
+  foregroundOpen: boolean;
+  foregroundToken: symbol;
+  retained?: ReturnType<typeof retainBeforeToolCallForNativeHookRelay>;
+  retention?: NativeHookRelayRetention;
+  removeAbortListener?: () => void;
+  expiryTimer?: ReturnType<typeof setTimeout>;
+};
+
+const RELAY_LIFETIME = "__openclawNativeHookRelayLifetimeV1";
+
+/** Private bundled-runtime callbacks for retained direct-child hook policy. */
+export type NativeHookRelayRetention = Readonly<{
+  readClaim: (rawPayload: unknown) => string | undefined;
+  shouldRetainAfterForegroundClose: () => boolean;
+  allowPreToolUse: (claim: string) => boolean;
+  awaitForegroundAdmission?: (claim: string) => Promise<(() => boolean) | undefined>;
+  onDispose: () => void;
+}>;
+
+type RetainedNativeHookRelayParams = RegisterNativeHookRelayParams & {
+  retention: NativeHookRelayRetention;
+};
+
+function readRelayLifetime(
+  registration: ActiveNativeHookRelayRegistration,
+): RelayLifetime | undefined {
+  // SAFETY: this private symbol-keyed expando is installed only by setRelayLifetime below.
+  return (registration as ActiveNativeHookRelayRegistration & { [RELAY_LIFETIME]?: RelayLifetime })[
+    RELAY_LIFETIME
+  ];
+}
+
+function setRelayLifetime(
+  registration: ActiveNativeHookRelayRegistration,
+  lifetime: RelayLifetime,
+): void {
+  Object.defineProperty(registration, RELAY_LIFETIME, {
+    configurable: true,
+    value: lifetime,
+  });
+}
+
+function scheduleNativeHookRelayExpiry(
+  relayId: string,
+  registration: ActiveNativeHookRelayRegistration,
+): void {
+  const lifetime = readRelayLifetime(registration);
+  if (!lifetime) {
+    return;
+  }
+  if (lifetime.expiryTimer) {
+    clearTimeout(lifetime.expiryTimer);
+  }
+  const rearm = () => {
+    if (relays.get(relayId) !== registration) {
+      return;
+    }
+    const remainingMs = registration.expiresAtMs - Date.now();
+    if (remainingMs < 0) {
+      unregisterNativeHookRelay(relayId, registration);
+      return;
+    }
+    lifetime.expiryTimer = setTimeout(rearm, Math.min(remainingMs + 1, MAX_TIMER_TIMEOUT_MS));
+    lifetime.expiryTimer.unref();
+  };
+  rearm();
+}
+
+function resolveNativeHookRelayExpiresAtMs(ttlMs: number | undefined): number | undefined {
+  return resolveExpiresAtMsFromDurationMs(normalizePositiveInteger(ttlMs, DEFAULT_RELAY_TTL_MS));
+}
 
 export function registerNativeHookRelay(
   params: RegisterNativeHookRelayParams,
 ): ActiveNativeHookRelayRegistrationHandle {
-  return registerNativeHookRelayLifecycle(params, invokeNativeHookRelay);
+  return registerNativeHookRelayInternal(params, undefined);
 }
 
-type RetainedNativeHookRelayParams = RegisterNativeHookRelayParams & {
-  composeWithExistingRoute?: boolean;
-  retention: NativeHookRelayRetention;
-};
+/** Private-local bundled runtime entrypoint; not exported through the public SDK. */
+export function registerRetainedNativeHookRelay(
+  params: RetainedNativeHookRelayParams,
+): ActiveNativeHookRelayRegistrationHandle {
+  const { retention, ...registrationParams } = params;
+  return registerNativeHookRelayInternal(registrationParams, retention);
+}
 
-export function registerRetainedNativeHookRelay(params: RetainedNativeHookRelayParams) {
-  return registerRetainedNativeHookRelayLifecycle(params, invokeNativeHookRelay);
+function registerNativeHookRelayInternal(
+  params: RegisterNativeHookRelayParams,
+  retention: NativeHookRelayRetention | undefined,
+): ActiveNativeHookRelayRegistrationHandle {
+  pruneExpiredNativeHookRelays();
+  pruneNativeHookRelayPermissionAllowAlways();
+  const relayId = normalizeRelayKey(params.relayId, "id") ?? randomUUID();
+  const generation = normalizeRelayKey(params.generation, "generation") ?? randomUUID();
+  const generationMismatchGraceMs = normalizePositiveInteger(params.generationMismatchGraceMs, 0);
+  const now = Date.now();
+  const expiresAtMs = resolveNativeHookRelayExpiresAtMs(params.ttlMs);
+  if (expiresAtMs === undefined) {
+    throw new Error("Native hook relay expiry is outside the supported Date range");
+  }
+  const allowedEvents = normalizeAllowedEvents(params.allowedEvents);
+  const stateDbPath = resolveOpenClawStateSqlitePath();
+  const deliverReplacedRegistrationUnregister = unregisterNativeHookRelay(relayId, undefined, {
+    deferBridgeRecordRemovalMs: NATIVE_HOOK_BRIDGE_REPLACEMENT_RECORD_GRACE_MS,
+    deferOnUnregister: true,
+  });
+  let partialRegistration: ActiveNativeHookRelayRegistration | undefined;
+  try {
+    const retained =
+      params.runBeforeToolCall && retention
+        ? retainBeforeToolCallForNativeHookRelay(params.runBeforeToolCall)
+        : undefined;
+    const registration = {
+      relayId,
+      provider: params.provider,
+      generation,
+      ...(generationMismatchGraceMs > 0
+        ? { generationMismatchGraceExpiresAtMs: now + generationMismatchGraceMs }
+        : {}),
+      ...(params.agentId ? { agentId: params.agentId } : {}),
+      sessionId: params.sessionId,
+      ...(params.sessionKey ? { sessionKey: params.sessionKey } : {}),
+      ...(params.config ? { config: params.config } : {}),
+      runId: params.runId,
+      ...(params.channelId ? { channelId: params.channelId } : {}),
+      ...(params.requester ? { requester: params.requester } : {}),
+      ...(params.approvalContext ? { approvalContext: params.approvalContext } : {}),
+      allowedEvents,
+      preToolUseLoopDetection: params.preToolUseLoopDetection !== false,
+      expiresAtMs,
+      preToolUseFailureProjections: new Map(),
+      ...(params.signal ? { signal: params.signal } : {}),
+      ...(params.runBeforeToolCall ? { runBeforeToolCall: params.runBeforeToolCall } : {}),
+      ...(params.assertActive ? { assertActive: params.assertActive } : {}),
+      ...(params.onPreToolUseFailure ? { onPreToolUseFailure: params.onPreToolUseFailure } : {}),
+      // SAFETY: the literal supplies the complete mutable internal registration contract.
+    } as ActiveNativeHookRelayRegistration;
+    partialRegistration = registration;
+    relays.set(relayId, registration);
+    setRelayLifetime(registration, {
+      foregroundOpen: true,
+      foregroundToken: Symbol("native-hook-relay-foreground"),
+      ...(retained ? { retained } : {}),
+      ...(retention ? { retention } : {}),
+    });
+    if (params.signal) {
+      const abort = () => unregisterNativeHookRelay(relayId, registration);
+      params.signal.addEventListener("abort", abort, { once: true });
+      readRelayLifetime(registration)!.removeAbortListener = () =>
+        params.signal?.removeEventListener("abort", abort);
+      if (params.signal.aborted) {
+        unregisterNativeHookRelay(relayId, registration);
+        throw new Error("native hook relay registration aborted");
+      }
+    }
+    registerNativeHookRelayBridge(registration, stateDbPath, invokeNativeHookRelay);
+    scheduleNativeHookRelayExpiry(relayId, registration);
+    const handle: ActiveNativeHookRelayRegistrationHandle = {
+      ...registration,
+      shouldRelayEvent: (event) => nativeHookRelayEventHasLocalWork(registration, event),
+      toolMatcherForEvent: (event) => nativeHookRelayEventToolMatcher(registration, event),
+      commandForEvent: (event, options) =>
+        buildNativeHookRelayCommandWithStateDatabase({
+          provider: params.provider,
+          relayId,
+          stateDbPath,
+          generation: registration.generation,
+          event,
+          nice: params.command?.nice,
+          timeoutMs: resolveNativeHookRelayCommandTimeoutMs(
+            params.command?.timeoutMs,
+            options?.timeoutMs,
+          ),
+          executable: params.command?.executable,
+          nodeExecutable: params.command?.nodeExecutable,
+        }),
+      renew: (ttlMs) => {
+        const current = relays.get(relayId);
+        if (current !== registration) {
+          return;
+        }
+        const renewedExpiresAtMs = resolveNativeHookRelayExpiresAtMs(ttlMs);
+        if (renewedExpiresAtMs === undefined) {
+          return;
+        }
+        const bridge = relayBridges.get(relayId);
+        if (bridge && bridge.server.listening) {
+          try {
+            const renewal = renewNativeHookRelayBridgeRecord(current, bridge, renewedExpiresAtMs);
+            if (renewal === "unavailable") {
+              return;
+            }
+            if (renewal === "ownership-changed") {
+              log.debug("native hook relay bridge record ownership changed", { relayId });
+              unregisterNativeHookRelay(relayId, current);
+              return;
+            }
+          } catch (error) {
+            log.debug("failed to renew native hook relay bridge record", { error, relayId });
+            return;
+          }
+        }
+        current.expiresAtMs = renewedExpiresAtMs;
+        handle.expiresAtMs = renewedExpiresAtMs;
+        scheduleNativeHookRelayExpiry(relayId, current);
+      },
+      unregister: () => deactivateNativeHookRelayForeground(relayId, registration),
+    };
+    return handle;
+  } catch (error) {
+    if (partialRegistration) {
+      unregisterNativeHookRelay(relayId, partialRegistration);
+    }
+    throw error;
+  } finally {
+    // The successor is authoritative before the old callback runs. A reentrant
+    // callback can therefore replace this registration normally instead of
+    // being overwritten by the outer replacement path. Finally also preserves
+    // the old callback if successor setup aborts partway through.
+    deliverReplacedRegistrationUnregister?.();
+  }
+}
+
+function unregisterNativeHookRelay(
+  relayId: string,
+  expectedRegistration?: ActiveNativeHookRelayRegistration,
+  options?: { deferBridgeRecordRemovalMs?: number; deferOnUnregister?: boolean },
+): (() => void) | undefined {
+  if (expectedRegistration && relays.get(relayId) !== expectedRegistration) {
+    return undefined;
+  }
+  const registration = expectedRegistration ?? relays.get(relayId);
+  if (!registration) {
+    return undefined;
+  }
+  const lifetime = readRelayLifetime(registration);
+  const bridge = relayBridges.get(relayId);
+  // Detach first: owner cleanup may register a same-id successor, which must
+  // never be removed by this registration's later resource cleanup.
+  if (relays.get(relayId) === registration) {
+    relays.delete(relayId);
+  }
+  if (lifetime?.expiryTimer) {
+    clearTimeout(lifetime.expiryTimer);
+  }
+  lifetime?.removeAbortListener?.();
+  lifetime?.retained?.release();
+  // SAFETY: this deletes the same private expando installed by setRelayLifetime.
+  delete (registration as ActiveNativeHookRelayRegistration & { [RELAY_LIFETIME]?: RelayLifetime })[
+    RELAY_LIFETIME
+  ];
+  unregisterNativeHookRelayBridge(relayId, {
+    ...options,
+    ...(bridge ? { expectedBridge: bridge } : {}),
+  });
+  removeNativeHookRelayInvocations(relayId);
+  removeNativeHookRelayPreToolUseApprovals(relayId);
+  removeNativeHookRelayPermissionState(relayId);
+  const deliverOnUnregister = () => {
+    try {
+      lifetime?.retention?.onDispose();
+    } catch (error) {
+      try {
+        log.warn("native hook relay unregister callback failed", { error, relayId });
+      } catch {
+        // Teardown has already detached every identity-bound resource. Logging
+        // must not turn an observer callback failure into a cleanup failure.
+      }
+    }
+  };
+  if (options?.deferOnUnregister) {
+    return deliverOnUnregister;
+  }
+  deliverOnUnregister();
+  return undefined;
+}
+
+function deactivateNativeHookRelayForeground(
+  relayId: string,
+  registration: ActiveNativeHookRelayRegistration,
+): void {
+  if (relays.get(relayId) !== registration) {
+    return;
+  }
+  const lifetime = readRelayLifetime(registration);
+  if (!lifetime) {
+    return;
+  }
+  lifetime.foregroundOpen = false;
+  let shouldRetain = false;
+  if (lifetime.retained && lifetime.retention) {
+    try {
+      shouldRetain = lifetime.retention.shouldRetainAfterForegroundClose();
+    } catch (error) {
+      try {
+        log.warn("native hook relay retention predicate failed", { error, relayId });
+      } catch {
+        // A logging failure cannot make a throwing retention predicate retain authority.
+      }
+    }
+  }
+  if (shouldRetain) {
+    return;
+  }
+  unregisterNativeHookRelay(relayId, registration);
+}
+
+async function resolveNativeHookRelayInvocationBinding(
+  registration: ActiveNativeHookRelayRegistration,
+  event: NativeHookRelayEvent,
+  rawPayload: unknown,
+): Promise<NativeHookRelayRegistration> {
+  const lifetime = readRelayLifetime(registration);
+  if (!lifetime) {
+    throw new Error("native hook relay registration is inactive");
+  }
+  const claim = lifetime.retention?.readClaim(rawPayload);
+  if (claim && event === "pre_tool_use" && lifetime.retained && lifetime.retention) {
+    const retained = lifetime.retained;
+    const retention = lifetime.retention;
+    let assertAdmission: (() => boolean) | undefined;
+    const assertRetainedAuthority = () => {
+      if (
+        relays.get(registration.relayId) !== registration ||
+        Date.now() > registration.expiresAtMs
+      ) {
+        throw new Error("native hook relay registration is inactive");
+      }
+      registration.signal?.throwIfAborted();
+      retained.assertActive();
+      if (assertAdmission && !assertAdmission()) {
+        throw new Error("native hook relay retained invocation not allowed");
+      }
+      if (!retention.allowPreToolUse(claim)) {
+        throw new Error("native hook relay retained invocation not allowed");
+      }
+    };
+    if (lifetime.foregroundOpen && retention.awaitForegroundAdmission) {
+      assertAdmission = await retention.awaitForegroundAdmission(claim);
+      if (!assertAdmission) {
+        throw new Error("native hook relay retained invocation not allowed");
+      }
+      assertRetainedAuthority();
+    } else if (!retention.allowPreToolUse(claim)) {
+      throw new Error("native hook relay retained invocation not allowed");
+    }
+    return {
+      ...registration,
+      assertActive: assertRetainedAuthority,
+      runBeforeToolCall: retained.runBeforeToolCall,
+    };
+  }
+  if (!lifetime.foregroundOpen) {
+    throw new Error("native hook relay foreground invocation not allowed");
+  }
+  const foregroundToken = lifetime.foregroundToken;
+  const assertActive = () => {
+    if (
+      relays.get(registration.relayId) !== registration ||
+      Date.now() > registration.expiresAtMs
+    ) {
+      throw new Error("native hook relay registration is inactive");
+    }
+    registration.signal?.throwIfAborted();
+    registration.assertActive?.();
+    if (!lifetime.foregroundOpen || lifetime.foregroundToken !== foregroundToken) {
+      throw new Error("native hook relay foreground invocation not allowed");
+    }
+  };
+  return { ...registration, assertActive };
+}
+
+function normalizeRelayKey(
+  value: string | undefined,
+  kind: "id" | "generation",
+): string | undefined {
+  const trimmed = value?.trim();
+  if (!trimmed) {
+    return undefined;
+  }
+  if (trimmed.length > 160 || !/^[A-Za-z0-9._:-]+$/u.test(trimmed)) {
+    throw new Error(`native hook relay ${kind} must be non-empty, compact, and URL-safe`);
+  }
+  return trimmed;
 }
 
 export async function invokeNativeHookRelay(
@@ -89,49 +475,51 @@ export async function invokeNativeHookRelay(
   const provider = readNativeHookRelayProvider(params.provider);
   const relayId = readNonEmptyString(params.relayId, "relayId");
   const event = readNativeHookRelayEvent(params.event);
-  const route = getNativeHookRelayRoute(relayId);
-  if (!route) {
+  const registration = relays.get(relayId);
+  if (!registration) {
     pruneExpiredNativeHookRelays();
     throw new Error("native hook relay not found");
   }
-  if (route.provider !== provider) {
-    throw new Error("native hook relay provider mismatch");
-  }
-  if (Date.now() > route.expiresAtMs) {
-    unregisterNativeHookRelay(relayId, route);
+  if (Date.now() > registration.expiresAtMs) {
+    unregisterNativeHookRelay(relayId, registration);
     throw new Error("native hook relay expired");
   }
-  if (!isJsonValue(params.rawPayload)) {
-    throw new Error("native hook relay payload must be JSON-compatible");
+  if (registration.provider !== provider) {
+    throw new Error("native hook relay provider mismatch");
   }
   if (params.requireGeneration) {
     const generation = readNonEmptyString(params.generation, "generation");
-    if (generation !== route.generation) {
-      if (!canAcceptNativeHookRelayGenerationMismatch(route, generation)) {
+    if (generation !== registration.generation) {
+      if (!canAcceptNativeHookRelayGenerationMismatch(registration, generation)) {
         throw new Error(NATIVE_HOOK_RELAY_BRIDGE_STALE_REGISTRATION_ERROR);
       }
       log.debug("native hook relay accepted bootstrap generation mismatch", {
         relayId,
         event,
-        runId: route.runId,
+        runId: registration.runId,
       });
     }
   }
-  const effectiveRegistration = await resolveNativeHookRelayInvocationBinding(
-    route,
-    event,
-    params.rawPayload,
-  );
-  if (!effectiveRegistration.allowedEvents.includes(event)) {
+  if (!registration.allowedEvents.includes(event)) {
     throw new Error("native hook relay event not allowed");
+  }
+  if (!isJsonValue(params.rawPayload)) {
+    throw new Error("native hook relay payload must be JSON-compatible");
   }
 
   const normalized = normalizeNativeHookInvocation({
-    registration: effectiveRegistration,
+    registration,
     event,
     rawPayload: params.rawPayload,
   });
-  effectiveRegistration.assertActive?.();
+  const effectiveRegistration = await resolveNativeHookRelayInvocationBinding(
+    registration,
+    event,
+    params.rawPayload,
+  );
+  if (event === "pre_tool_use" || event === "permission_request") {
+    effectiveRegistration.assertActive?.();
+  }
   recordNativeHookRelayInvocation(normalized);
   const startedAt = Date.now();
   const response = await processNativeHookRelayInvocation({
@@ -141,11 +529,7 @@ export async function invokeNativeHookRelay(
   });
   // Policy and approval callbacks may yield while their admitted run closes.
   // Never let a late allow cross back into the native runtime.
-  if (
-    event === "pre_tool_use" ||
-    event === "permission_request" ||
-    event === "before_agent_finalize"
-  ) {
+  if (event === "pre_tool_use" || event === "permission_request") {
     effectiveRegistration.assertActive?.();
   }
   if (
@@ -153,7 +537,7 @@ export async function invokeNativeHookRelay(
     response.failureDisposition &&
     readNativeHookRelayApprovalMode(normalized.rawPayload) !== "report"
   ) {
-    projectNativeHookRelayPreToolUseFailure(effectiveRegistration, {
+    projectNativeHookRelayPreToolUseFailure(registration, {
       toolName: normalizeNativeHookToolName(normalized.toolName),
       toolCallId: normalized.toolUseId,
       disposition: response.failureDisposition,
@@ -210,19 +594,16 @@ function projectNativeHookRelayPreToolUseFailure(
 export function hasNativeHookRelayInvocation(params: {
   relayId: string;
   event: NativeHookRelayEvent;
-  turnId?: string;
   toolUseId?: string;
 }): boolean {
-  const turnId = params.turnId?.trim();
   const toolUseId = params.toolUseId?.trim();
-  if (!turnId || !toolUseId) {
+  if (!toolUseId) {
     return false;
   }
   return invocations.some(
     (invocation) =>
       invocation.relayId === params.relayId &&
       invocation.event === params.event &&
-      invocation.turnId === turnId &&
       invocation.toolUseId === toolUseId,
   );
 }
@@ -234,6 +615,14 @@ function recordNativeHookRelayInvocation(invocation: NativeHookRelayInvocation):
   });
   if (invocations.length > MAX_NATIVE_HOOK_RELAY_INVOCATIONS) {
     invocations.splice(0, invocations.length - MAX_NATIVE_HOOK_RELAY_INVOCATIONS);
+  }
+}
+
+function removeNativeHookRelayInvocations(relayId: string): void {
+  for (let index = invocations.length - 1; index >= 0; index -= 1) {
+    if (invocations[index]?.relayId === relayId) {
+      invocations.splice(index, 1);
+    }
   }
 }
 
@@ -252,9 +641,26 @@ function canAcceptNativeHookRelayGenerationMismatch(
   return true;
 }
 
+function pruneExpiredNativeHookRelays(now = Date.now()): void {
+  for (const [relayId, registration] of relays) {
+    if (now > registration.expiresAtMs) {
+      unregisterNativeHookRelay(relayId, registration);
+    }
+  }
+}
+
+function normalizeAllowedEvents(
+  events: readonly NativeHookRelayEvent[] | undefined,
+): readonly NativeHookRelayEvent[] {
+  if (!events?.length) {
+    return NATIVE_HOOK_RELAY_EVENTS;
+  }
+  return [...new Set(events)];
+}
+
 export const testing = {
   clearNativeHookRelaysForTests(): void {
-    for (const [relayId, registration] of listNativeHookRelayRoutes()) {
+    for (const [relayId, registration] of relays) {
       unregisterNativeHookRelay(relayId, registration);
     }
     clearNativeHookRelayBridgesForTests();
@@ -265,7 +671,7 @@ export const testing = {
     return [...invocations];
   },
   getNativeHookRelayRegistrationForTests(relayId: string): NativeHookRelayRegistration | undefined {
-    return getNativeHookRelayRegistrationForTests(relayId);
+    return relays.get(relayId);
   },
   getNativeHookRelayBridgeDirForTests(): string {
     throw new Error("native hook relay bridge files were retired");
@@ -281,9 +687,18 @@ export const testing = {
   isNativeHookRelayBridgeLookupRetryableForTests(error: unknown, elapsedMs = 0): boolean {
     return isRetryableNativeHookRelayBridgeLookupError({ error, elapsedMs });
   },
-  formatPermissionApprovalDescriptionForTests: formatPermissionApprovalDescription,
-  permissionRequestContentFingerprintForTests: permissionRequestContentFingerprint,
-  permissionRequestToolInputKeyFingerprintForTests: permissionRequestToolInputKeyFingerprint,
+  formatPermissionApprovalDescriptionForTests(
+    request: NativeHookRelayPermissionApprovalRequest,
+  ): string {
+    return formatPermissionApprovalDescriptionForTestsImpl(request);
+  },
+  permissionRequestContentFingerprintForTests(
+    request: NativeHookRelayPermissionApprovalRequest,
+  ): string {
+    return permissionRequestContentFingerprintForTestsImpl(request);
+  },
+  permissionRequestToolInputKeyFingerprintForTests:
+    permissionRequestToolInputKeyFingerprintForTestsImpl,
   setNativeHookRelayPermissionApprovalRequesterForTests(
     requester: NativeHookRelayPermissionApprovalRequester,
   ): void {
